requirements.txt
absl-py==1.0.0
attrs==21.4.0
click==8.0.4
cmake==3.22.2
comtypes==1.1.11
cycler==0.11.0
# dlib==19.23.0
enum34==1.1.10
evdev==1.4.0
face-recognition==1.3.0
face-recognition-models==0.3.0
fonttools==4.29.1
future==0.18.2
keyboard==0.13.5
kiwisolver==1.3.2
matplotlib==3.5.1
mediapipe==0.8.9.1
MouseInfo==0.1.3
numpy==1.22.2
opencv-contrib-python==4.5.5.62
opencv-python==4.5.5.62
packaging==21.3
Pillow==9.0.1
protobuf==3.19.4
psutil==5.9.0
PyAutoGUI==0.9.53
pycaw==20181226
pyCoW==0.0.4
PyGetWindow==0.0.9
pymongo==4.0.1
PyMsgBox==1.0.9
pynput==1.7.6
pyparsing==3.0.7
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.28
python-dateutil==2.8.2
python-xlib==0.31
python3-xlib==0.15
pytweening==1.0.4
semantic-version==2.9.0
setuptools-rust==1.1.2
six==1.16.0
tk==0.1.0
typing_extensions==4.0.1
